+++ app/src/main/java/abandonedstudio/app/tospace/core/domain/model/DetailedLaunch.kt
@@ -19,7 +19,7 @@ data class DetailedLaunch(
 
     data class Payload(
         val type: String?,
-        val massKg: Int?,
+        val massKg: Double?,
         val orbit: String?,
         val inclination: Double?,
         val periodMin: Double?,
